requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==5.0.3
DateTime==4.3
Flask==1.0.2
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mock==4.0.1
nose==1.3.7
nose2==0.9.2
prometheus-client==0.7.1
pyArango==1.3.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
zope.interface==4.7.1
